feat: main 구현
